requirements.txt
# Google Document AI
google-cloud-documentai>=2.20.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# LLM APIs
google-genai>=0.3.0
openai>=1.68.0

# Data processing
pandas>=2.0.0

# PDF handling
PyPDF2>=3.0.0

# HTTP requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

